Update version and refactor event handling
The version number in pyproject.toml has been incremented. Also, the handling of instances in the state machine has been adjusted to utilize the Model's __all_instances__ attribute. Additionally, changes have been implemented in the traffic_light.py example for more streamlined handling of the FlashingEvent in the TrafficLight class, as well as adjustments in the implementation of the Signal submachine state in the Pedestrian class.